requirements.txt
Django
pylint
topicaxis-opengraph